Use Walmart's App Store icon.
Summary:
This fixes an issue where Walmart's icon is not showing up in production. I added the icon to the RN website, but this goes against convention which is to use an externally hosted image for the showcase.

This PR uses the App Store icon for the app.

Before:

<img width="1069" alt="unknown" src="https://cloud.githubusercontent.com/assets/165856/21729495/3bb95990-d400-11e6-98b9-9ffd86a1b3b6.png">

After:
![screen shot 2017-01-06 at 11 01 31 am](https://cloud.githubusercontent.com/assets/165856/21729508/44d049bc-d400-11e6-81ff-87c9550a8ea6.png)
Closes https://github.com/facebook/react-native/pull/11754

Differential Revision: D4387761

Pulled By: ericvicenti

fbshipit-source-id: 1be4903c1f05be2cfd8403f8746c1917669d196f