disable Circle e2e tests
Summary:
Motivation is that Circle e2e tests have been broken for a few days, it appears to be flaky and an ongoing problem. AaaChiuuu & Andrew Chen are taking a look, but for now let's turn this off so that Circle can be useful for other things.
Closes https://github.com/facebook/react-native/pull/13616

Differential Revision: D4931852

Pulled By: ericvicenti

fbshipit-source-id: ce725ff3311ea70052191df45d3feb4373176876